Refactor server module
